title updates and calories calculations

+++ Mobile/lib/bar_graphs/individualbar.dart
@@ -1,5 +1,5 @@
 class individualBar {
-  int x;
+  double x;
   double y;
   //String x;
